+++ jsonargparse_tests/base.py
@@ -14,7 +14,7 @@ from jsonargparse.optionals import (
     dataclasses_support, import_dataclasses,
     fsspec_support, import_fsspec,
     get_config_read_mode,
-    ruyaml_support,
+    ruyaml_support, import_ruyaml,
     ModuleNotFound,
 )