+++ CHANGELOG.rst
@@ -10,6 +10,18 @@ only be introduced in major versions with advance notice in the **Deprecated**
 section of releases.
 
 
+v3.13.0 (2021-??-??)
+--------------------
+
+Added
+^^^^^
+- Inner config file support for subclass type hints in signatures and CLI #57.
+
+Fixed
+^^^^^
+- With fail_untyped=True use type from default value instead of Any.
+
+
 v3.12.1 (2021-05-19)
 --------------------